+++ boosters/node-beginner/meeting-3/tests/03_exercise.test.js
@@ -1,20 +1,33 @@
 const request = require("supertest");
 const app = require("../src/server");
+const setup = require("./config/setup");
 const teardown = require("./config/teardown");
 const uuid = require("uuid/v4");
 
-const user = {
-  firstName: "Andrew",
-  lastName: "Mayer",
-  id: uuid(),
-};
-describe("Create User", () => {
-  afterEach(teardown);
+const users = [
+  { firstName: "Andrew", lastName: "Mayer", id: uuid() },
+  { firstName: "Liang", lastName: "Chen", id: uuid() },
+];
 
-  xit("Should create a user", async () => {
-    const res = await request(app)
-      .post("/users/")
-      .send(user);
-    expect(res.statusCode).toEqual(201);
+const repos = users.reduce(
+  (acc, curr) => [
+    ...acc,
+    { name: "repo1", userId: curr.id, isPrivate: false, id: uuid() },
+    { name: "repo2", userId: curr.id, isPrivate: true, id: uuid() },
+  ],
+  []
+);
+
+describe("Repos", () => {
+  beforeEach(async () => {
+    await setup("Users", users)();
+    return setup("Repos", repos)();
+  });
+  afterEach(teardown);
+  xit("Should return all repos", async () => {
+    res = await request(app).get(`/repos`);
+
+    expect(res.statusCode).toEqual(200);
+    expect(res.body.length).toEqual(repos.length);
   });
 });
