+++ boosters/node-beginner/meeting-4/src/api/controllers/userController.js
@@ -28,15 +28,4 @@ const getUser = async (req, res) => {
   res.send(user);
 };
 
-const getUsersRepo = async (req, res) => {
-  const { userId } = req.params;
-
-  const repos = await Repos.findAll({
-    where: {
-      userId,
-    },
-  });
-  res.send(repos);
-};
-
-module.exports = { getAllUsers, createUser, getUser, getUsersRepo };
+module.exports = { getAllUsers, createUser, getUser };
