+++ boosters/node-beginner/meeting-4/src/api/routes/repos/index.js
@@ -2,11 +2,6 @@ const express = require("express");
 const router = express.Router();
 const { repoController } = require("../../controllers");
 
-router
-  .route("/")
-  .get(repoController.getAllRepos)
-  .post(repoController.createRepo);
-
-router.get("/:repoId", repoController.getRepo);
+router.route("/").get(repoController.getAllRepos);
 
 module.exports = router;
