+++ boosters/node-beginner/meeting-4/tests/00_exercise.test.js
@@ -9,12 +9,27 @@ const users = [
   { firstName: "Liang", lastName: "Chen", id: uuid() },
 ];
 
-describe("Get All Users", () => {
-  beforeEach(setup("Users", users));
+const repos = users.reduce(
+  (acc, curr) => [
+    ...acc,
+    { name: "repo1", userId: curr.id, isPrivate: false, id: uuid() },
+    { name: "repo2", userId: curr.id, isPrivate: true, id: uuid() },
+  ],
+  []
+);
+
+describe("User Repos", () => {
+  beforeEach(async () => {
+    await setup("Users", users)();
+    return setup("Repos", repos)();
+  });
   afterEach(teardown);
-  xit("Should return all users", async () => {
-    const res = await request(app).get("/users/");
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.length).toEqual(users.length);
+  xit("Should return all repos of a given userId", async () => {
+    res = await Promise.all(users.map(user => request(app).get(`/users/${user.id}/repos`)));
+
+    res.map(res => {
+      expect(res.statusCode).toEqual(200);
+      expect(res.body.length).toEqual(2);
+    });
   });
 });
