+++ boosters/node-beginner/meeting-4/tests/01_exercise.test.js
@@ -1,25 +1,31 @@
 const request = require("supertest");
 const app = require("../src/server");
-const setup = require("./config/setup");
 const teardown = require("./config/teardown");
 const uuid = require("uuid/v4");
+const setup = require("./config/setup");
 
-const users = [
-  { firstName: "Andrew", lastName: "Mayer", id: uuid() },
-  { firstName: "Liang", lastName: "Chen", id: uuid() },
+const user = [
+  {
+    firstName: "Andrew",
+    lastName: "Mayer",
+    id: uuid(),
+  },
 ];
 
-describe("Specific user", () => {
-  beforeEach(setup("Users", users));
-  afterEach(teardown);
+const repo = {
+  name: "repo1",
+  userId: user.id,
+  isPrivate: false,
+};
 
-  xit("Should return user profile given speicifc userId", async () => {
-    res = await Promise.all(users.map(user => request(app).get(`/users/${user.id}`)));
+describe("Create Repo", () => {
+  beforeEach(setup("Users", user));
+  afterAll(teardown);
 
-    users.map((user, i) => {
-      expect(res[i].statusCode).toEqual(200);
-      expect(res[i].body.firstName).toEqual(user.firstName);
-      expect(res[i].body.lastName).toEqual(user.lastName);
-    });
+  xit("Should create a repo", async () => {
+    const res = await request(app)
+      .post("/repos/")
+      .send(repo);
+    expect(res.statusCode).toEqual(201);
   });
 });
